+++ cmd/terramate/cli/cloud.go
@@ -54,10 +54,10 @@ type cloudRunState struct {
 
 	stackMeta2ID map[string]int64
 	// stackPreviews is a map of stack.ID to stackPreview.ID
-	stackPreviews  map[string]string
-	reviewRequest  *cloud.ReviewRequest
-	prFromGHAEvent *github.PullRequest
-	metadata       *cloud.DeploymentMetadata
+	stackMeta2PreviewIDs map[string]string
+	reviewRequest        *cloud.ReviewRequest
+	prFromGHAEvent       *github.PullRequest
+	metadata             *cloud.DeploymentMetadata
 }
 
 type cloudConfig struct {
@@ -96,6 +96,18 @@ func (rs *cloudRunState) setMeta2CloudID(metaID string, id int64) {
 
 func (rs cloudRunState) stackCloudID(metaID string) (int64, bool) {
 	id, ok := rs.stackMeta2ID[strings.ToLower(metaID)]
+	return id, ok
+}
+
+func (rs *cloudRunState) setMeta2PreviewID(metaID string, previewID string) {
+	if rs.stackMeta2PreviewIDs == nil {
+		rs.stackMeta2PreviewIDs = make(map[string]string)
+	}
+	rs.stackMeta2PreviewIDs[strings.ToLower(metaID)] = previewID
+}
+
+func (rs cloudRunState) cloudPreviewID(metaID string) (string, bool) {
+	id, ok := rs.stackMeta2PreviewIDs[strings.ToLower(metaID)]
 	return id, ok
 }
 
@@ -163,10 +175,6 @@ func (c *cli) checkCloudSync() {
 		uuid, err := uuid.GenerateUUID()
 		c.handleCriticalError(err)
 		c.cloud.run.runUUID = cloud.UUID(uuid)
-	}
-
-	if c.parsedArgs.Run.CloudSyncPreview {
-		c.cloud.run.stackPreviews = make(map[string]string)
 	}
 }
 
@@ -308,7 +316,11 @@ func (c *cli) cloudSyncAfter(run stackCloudRun, res runResult, err error) {
 }
 
 func (c *cli) doPreviewBefore(run stackCloudRun) {
-	stackPreviewID := c.cloud.run.stackPreviews[run.Stack.ID]
+	stackPreviewID, ok := c.cloud.run.cloudPreviewID(run.Stack.ID)
+	if !ok {
+		c.disableCloudFeatures(errors.E(errors.ErrInternal, "failed to get previewID"))
+		return
+	}
 	ctx, cancel := context.WithTimeout(context.Background(), defaultCloudTimeout)
 	defer cancel()
 	if err := c.cloud.client.UpdateStackPreview(ctx,
@@ -355,7 +367,11 @@ func (c *cli) doPreviewAfter(run stackCloudRun, res runResult) {
 		}
 	}
 
-	stackPreviewID := c.cloud.run.stackPreviews[run.Stack.ID]
+	stackPreviewID, ok := c.cloud.run.cloudPreviewID(run.Stack.ID)
+	if !ok {
+		c.disableCloudFeatures(errors.E(errors.ErrInternal, "failed to get previewID"))
+		return
+	}
 	ctx, cancel := context.WithTimeout(context.Background(), defaultCloudTimeout)
 	defer cancel()
 	if err := c.cloud.client.UpdateStackPreview(ctx,